Add device id for disambiguation
When installing a lot of z-wave devices all at once, the preference
panel shows up with a list of "Switch" devices that makes it hard to
figure out which switch is which. Adding the device id in the config
screen makes it easier to tell which device is which.

Also cleaned up a little of using a closure context variable in cases
where it was not necessary.