Fix dashboard showing all the widgets
There's a change in the method signature of "showingControl" that it
only takes 2 parameters now instead of 3. As a result, the model came
in as "forceView", which results in force viewing to be true.

Since the method signature changed at a later date, the function call
is changed to match.