Make switch button title editable on dashboard
This uses HTML5 for making the title of the switch editable. When
entering the "Rearrange and Settings" mode on the Dashboard, the
switch title becomes editable. On blur, the title is saved. There's
special handling of enter keys, since it does not make sense to have
newlines in the names, enter keys are interpreted as "saving."