Display device location in notifications and config
Currently it is impossible to distinguish a device called "Motion" in the room lounge from a device called "Motion" in the kitchen, forcing devices to be called "Lounge Motion", now in the room lounge I am forced to devices named "Lounge Motion", "Lounge Main Light" etc. This clutters the UI with redundant information.

This updates the device drop downs and the notifications to include a location if one is set.

I appreciate if users already have a Device in there lounge called "Lounge Motion" they will now in places see a device called "Lounge - Lounge Motion", but this is no worse that what they currently see when looking at devices in the room lounge.